pyproject.toml
[tool.poetry]
name = "wagtail-links"
version = "2.8.1"
description = "Wagtail links provides a consistent way to refer to links in a wagtail page."
authors = ["thelab <thelabdev@thelab.co>"]
readme = "README.md"
homepage = "https://gitlab.com/thelabnyc/wagtail-links"
repository = "https://gitlab.com/thelabnyc/wagtail-links"
license = "ISC"

[[tool.poetry.packages]]
include = "wagtail_links"
from = "src"

[tool.poetry.dependencies]
python = "^3.10"
wagtail = ">=5.2.5,<6.3"
django-stubs-ext = "^5.0.4"

[tool.poetry.group.dev.dependencies]
flake8 = ">=7.1.1"
tox = ">=4.21.2"
mypy = "^1.11.2"
django-stubs = "^5.0.4"
coverage = "^7.6.1"

[build-system]
requires = [ "poetry-core>=1.9.0",]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
from_first = true

[tool.commitizen]
name = "cz_conventional_commits"
annotated_tag = true
gpg_sign = true
tag_format = "v$version"
update_changelog_on_bump = true
changelog_merge_prerelease = true
version_provider = "poetry"
version_scheme = "pep440"
version_files = [
    "pyproject.toml:version",
]
pre_bump_hooks = [
    "pre-commit run --all-files || true",
]
post_bump_hooks = [
    "git push origin master $CZ_POST_CURRENT_TAG_VERSION"
]
